🚚 refactor: rename filter and search bar to have project prefix
