+++ src/app/page.tsx
@@ -8,7 +8,7 @@ import { TimeframeSearch } from '@/components/molecules/TimeframeSearch/Timefram
 import GeoMap from '@/components/organisms/GeoMap/GeoMap'
 import CreditsHistorySection from '@/components/organisms/CreditsHistorySection/CreditsHistorySection'
 import { useTranslation } from 'react-i18next'
-import FiltersAndSearch from '@/components/organisms/FiltersAndSearch/FiltersAndSearch'
+import ProjectFilterAndSearch from '@/components/organisms/ProjectFilterAndSearch/ProjectFilterAndSearch'
 
 export default function Home(): React.JSX.Element {
   const { carbonReduction } = useAppState().analytics
@@ -20,7 +20,7 @@ export default function Home(): React.JSX.Element {
   return (
     <>
       <Box paddingX={`24px`} paddingTop={`24px`}>
-        <FiltersAndSearch />
+        <ProjectFilterAndSearch />
       </Box>
       <Flex padding={`24px`} minHeight="min-content" alignItems="center" justifyContent="space-between" flexDirection="column" width={`100%`}>
         <Container marginTop={`20px`} variant="cardSection">
