+++ src/components/organisms/ProjectSearchBar/ProjectSearchBar.spec.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import SearchBar from '@/components/organisms/SearchBar/SearchBar'
+import ProjectSearchBar from '@/components/organisms/ProjectSearchBar/ProjectSearchBar'
 
 const mockPush = jest.fn()
 jest.mock(`next/navigation`, () => ({
@@ -10,12 +10,12 @@ jest.mock(`next/navigation`, () => ({
 }))
 
 it(`renders correctly`, () => {
-  const { container } = render(<SearchBar />)
+  const { container } = render(<ProjectSearchBar />)
   expect(container).toMatchSnapshot()
 })
 
 it(`renders correctly with no params, add a search value and submits a search`, async () => {
-  const { container } = render(<SearchBar />)
+  const { container } = render(<ProjectSearchBar />)
   await userEvent.type(screen.getByRole(`textbox`), `hello`)
   await userEvent.click(screen.getByTestId(`search-bar-clear`))
   await userEvent.click(screen.getByTestId(`search-bar-search`))
